updates to add more debugging into
git-svn-id: https://www.brinkman.mbb.sfu.ca/svn/psort/apache-svm/trunk@8 f6aa61ce-f74f-0410-b74d-074a105d0ac4

+++ binding.h
@@ -19,7 +19,7 @@ extern "C" {
   extern void destroySVM(SVM*);
   extern int loadSVMModel(SVM*, char*);
   extern int loadSVMFreqPattern(SVM*, char*);
-  extern int SVMClassify(SVM*, char*);
+  extern double SVMClassify(SVM*, char*);
 
 #ifdef __cplusplus
 }
